+++ projects/hell/src/lib/public-api-core.ts
@@ -2,5 +2,4 @@ export * from './core/types';
 export * from './core/styleable';
 export * from './core/search';
 export * from './core/labels';
-export * from './core/hotkeys';
 export * from './core/floating-element';